Junkyu

+++ backend/src/main/java/com/example/demo/dao/CategoryService.java
@@ -1,8 +1,7 @@
 package com.example.demo.dao;
 
-import com.example.demo.vo.CategoryVO;
+import com.example.demo.vo.Category.CategoryVO;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -10,17 +9,40 @@ import java.util.List;
 @Service
 public class CategoryService {
 
-    @Autowired(required = true)
-    private  CategoryRepository categoryRepository;
+    private final CategoryRepository categoryRepository;
+
+    @Autowired
+    public CategoryService(CategoryRepository categoryRepository){
+        this.categoryRepository=categoryRepository;
+    }
+
+    public void AddCategory(CategoryVO categoryVO) throws Exception{
+        try{
+            categoryRepository.AddCategory(categoryVO);
+        }catch(Exception e){
+            e.printStackTrace();
+            throw new Exception("INTERVAL SERVER ERROR");
+        }
 
 
-    public void AddCategory(CategoryVO categoryVO){
-        categoryRepository.AddCategory(categoryVO);
     }
 
 
-    public void DeleteCategory(Long categoryIdx){
-        categoryRepository.DeleteCategory(categoryIdx);
+    public CategoryVO DeleteCategory(Long categoryIdx) throws IndexOutOfBoundsException,Exception{
+        CategoryVO findCategoryVO = null;
+        try{
+            //먼저 삭제할 것이 있는지 확인해주고
+            findCategoryVO = categoryRepository.findCategoryByIdx(categoryIdx).get(0);
+            if(findCategoryVO != null){
+                categoryRepository.DeleteCategory(categoryIdx);
+            }
+        }catch(IndexOutOfBoundsException e){
+            //없다면 null처리
+            throw new IndexOutOfBoundsException("NO DATA BAD REQUEST");
+        }catch(Exception e){
+            throw new Exception("INTERVAL SERVER ERROR");
+        }
+        return findCategoryVO;
     }
 
 
@@ -30,6 +52,10 @@ public class CategoryService {
 
 
     public List<CategoryVO> getCategoryList(Long memberIdx){
-        return categoryRepository.findCategoryVOByMemberidx(memberIdx);
+        return categoryRepository.GetCategoryList(memberIdx);
+    }
+
+    public List<CategoryVO> findCategoryByIdx(Long idx){
+        return categoryRepository.findCategoryByIdx(idx);
     }
 }
